@@ -32,7 +32,6 @@ using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.DataProtection;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Infrastructure;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
@@ -40,6 +39,12 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.IdentityModel.Tokens;
 
 using Swashbuckle.AspNetCore.Swagger;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using Microsoft.AspNetCore.DataProtection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.OpenApi.Models;
 
 namespace Exoft.Gamification
 {
@@ -59,16 +64,17 @@ namespace Exoft.Gamification
                 .PersistKeysToFileSystem(new DirectoryInfo(@".\server\share\"));
 
             services.AddCors();
-            services.AddMvc(options =>
+            services.AddControllers(options =>
             {
+                options.EnableEndpointRouting = true;
                 options.Filters.Add<ErrorHandlingFilter>();
             })
-                .AddDataAnnotationsLocalization(options =>
-                {
-                    options.DataAnnotationLocalizerProvider = (type, factory) =>
-                        factory.Create(typeof(ValidatorMessages));
-                })
-                .SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+            .AddDataAnnotationsLocalization(options =>
+            {
+                options.DataAnnotationLocalizerProvider = (type, factory) =>
+                    factory.Create(typeof(ValidatorMessages));
+            })
+            .AddNewtonsoftJson();
 
             services.AddDbContext<UsersDbContext>
             (
@@ -169,26 +175,37 @@ namespace Exoft.Gamification
             // Swagger configuration
             services.AddSwaggerGen(c =>
             {
-                c.SwaggerDoc("v1", new Info { Title = "Gamification", Version = "0.0.0.1" });
+                c.SwaggerDoc("v1", new OpenApiInfo { Title = "Gamification", Version = "0.0.0.1" });
 
-                var security = new Dictionary<string, IEnumerable<string>>
-                {
-                    { "Bearer", new string[] { } }
-                };
-
-                c.AddSecurityDefinition("Bearer", new ApiKeyScheme
+                c.AddSecurityDefinition("Bearer", new OpenApiSecurityScheme
                 {
                     Description = "Example: \"Bearer {token}\"",
                     Name = "Authorization",
-                    In = "header",
-                    Type = "apiKey"
+                    In = ParameterLocation.Header,
+                    Type = SecuritySchemeType.ApiKey
                 });
-                c.AddSecurityRequirement(security);
+                c.AddSecurityRequirement(new OpenApiSecurityRequirement()
+                {
+                    {
+                        new OpenApiSecurityScheme
+                        {
+                        Reference = new OpenApiReference
+                            {
+                            Type = ReferenceType.SecurityScheme,
+                            Id = "Bearer"
+                            },
+                            Scheme = "oauth2",
+                            Name = "Bearer",
+                            In = ParameterLocation.Header,
+                        },
+                        new List<string>()
+                    }
+                });
             });
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, IServiceProvider serviceProvider)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, IServiceProvider serviceProvider)
         {
             using (var scope = app.ApplicationServices.GetRequiredService<IServiceScopeFactory>().CreateScope())
             {
@@ -209,6 +226,7 @@ namespace Exoft.Gamification
                 c.SwaggerEndpoint("/swagger/v1/swagger.json", "Gamification.Api");
             });
 
+            app.UseRouting();
             // global cors policy
             app.UseCors(x => x
                 .AllowAnyOrigin()
@@ -216,8 +234,11 @@ namespace Exoft.Gamification
                 .AllowAnyHeader());
 
             app.UseAuthentication();
+            app.UseAuthorization();
 
-            app.UseMvc();
+            app.UseEndpoints(endpoints => {
+                endpoints.MapControllers();
+            });
         }
     }
 }
